Split out scenario code
